Add body/main sections
Added section wrapper (left column) to include
a fun animation on load.

The right column includes more information
about the application and a button with the
ability to sign up for testing. The right
column also has an animation, which fires on
load.

Added the floc name to the navbar and made it
slide in on load as well.

Too many animation? I'm not sure.